# backend/study-buddy-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Study Buddy API
  description: |
    Production-grade backend for Study Buddy — an emotion-aware AI tutoring platform.
    
    ## Authentication
    All authenticated endpoints require a valid session cookie (`session`).
    Use `POST /api/v1/auth/login` or `POST /api/v1/auth/register` to obtain a session.
    The session is set as an `httpOnly` cookie automatically.
    
    ## Streaming (SSE)
    The `GET /api/v1/tutor/threads/{id}/stream` endpoint uses Server-Sent Events.
    Events emitted:
    - `token`: a text chunk from the AI assistant (`data` is a JSON string)
    - `done`: stream complete (`data` is `{"messageId":"..."}`)
    - `error`: an error occurred (`data` is `{"message":"..."}`)
  version: 1.0.0
  contact:
    name: Study Buddy Team

servers:
  - url: http://localhost:4000
    description: Local development

tags:
  - name: Auth
    description: Registration, login, logout, current user
  - name: Consent
    description: Privacy consent settings
  - name: StudySessions
    description: Study session management
  - name: Tutor
    description: AI tutor threads and messages
  - name: Emotions
    description: Emotion sample ingestion
  - name: Flashcards
    description: Flashcard decks and cards
  - name: Health
    description: Health check

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      operationId: healthCheck
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time
                  uptime:
                    type: number
        "503":
          description: Service is unhealthy

  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created and session cookie set
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email already registered
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in, session cookie set
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout (clear session)
      operationId: logout
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Logged out
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/v1/me:
    get:
      tags: [Auth]
      summary: Get current user
      operationId: getMe
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/v1/consent:
    get:
      tags: [Consent]
      summary: Get consent settings
      operationId: getConsent
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current consent settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentResponse"
    put:
      tags: [Consent]
      summary: Update consent settings
      operationId: updateConsent
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentInput"
      responses:
        "200":
          description: Updated consent settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentResponse"

  /api/v1/study-sessions:
    post:
      tags: [StudySessions]
      summary: Create a study session
      operationId: createStudySession
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudySessionInput"
      responses:
        "201":
          description: Study session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySessionResponse"

  /api/v1/tutor/threads:
    post:
      tags: [Tutor]
      summary: Create a tutor thread
      operationId: createThread
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateThreadInput"
      responses:
        "201":
          description: Thread created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreadResponse"
    get:
      tags: [Tutor]
      summary: List tutor threads
      operationId: listThreads
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of threads
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ThreadSummary"

  /api/v1/tutor/threads/{id}/messages:
    get:
      tags: [Tutor]
      summary: Get messages in a thread
      operationId: getThreadMessages
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Thread messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Message"
    post:
      tags: [Tutor]
      summary: Post a user message to a thread
      operationId: postMessage
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageInput"
      responses:
        "201":
          description: Message created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Message"

  /api/v1/tutor/threads/{id}/stream:
    get:
      tags: [Tutor]
      summary: Stream assistant response (SSE)
      description: |
        Server-Sent Events endpoint. After posting a user message, 
        call this endpoint to stream the AI assistant's response.
        
        **Events:**
        - `token` — partial text chunk (data is a JSON-encoded string)
        - `done` — stream complete (data is `{"messageId":"uuid"}`)
        - `error` — error occurred (data is `{"message":"..."}`)
      operationId: streamAssistant
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: afterMessageId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string

  /api/v1/emotions/samples:
    post:
      tags: [Emotions]
      summary: Submit emotion samples (batch)
      operationId: postEmotionSamples
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostEmotionSamplesInput"
      responses:
        "201":
          description: Samples ingested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmotionSamplesResponse"
        "403":
          description: Consent not granted

  /api/v1/flashcards/decks:
    post:
      tags: [Flashcards]
      summary: Create a flashcard deck
      operationId: createDeck
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeckInput"
      responses:
        "201":
          description: Deck created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeckResponse"
    get:
      tags: [Flashcards]
      summary: List flashcard decks
      operationId: listDecks
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of decks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/DeckSummary"

  /api/v1/flashcards/decks/{deckId}/cards:
    post:
      tags: [Flashcards]
      summary: Add a card to a deck
      operationId: createCard
      security:
        - cookieAuth: []
      parameters:
        - name: deckId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardInput"
      responses:
        "201":
          description: Card created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
    get:
      tags: [Flashcards]
      summary: List cards in a deck
      operationId: listCards
      security:
        - cookieAuth: []
      parameters:
        - name: deckId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CardResponse"

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session

  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        requestId:
          type: string

    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 10
        displayName:
          type: string

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    UserResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
            createdAt:
              type: string
              format: date-time

    MeResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
            createdAt:
              type: string
              format: date-time
            profile:
              type: object
              nullable: true
              properties:
                displayName:
                  type: string

    ConsentInput:
      type: object
      properties:
        emotionTelemetryAllowed:
          type: boolean
        webcamAllowed:
          type: boolean
        retentionDays:
          type: integer
          minimum: 1
          maximum: 365

    ConsentResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            emotionTelemetryAllowed:
              type: boolean
            webcamAllowed:
              type: boolean
            retentionDays:
              type: integer
            updatedAt:
              type: string
              format: date-time

    CreateStudySessionInput:
      type: object
      properties:
        topic:
          type: string

    StudySessionResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            topic:
              type: string
              nullable: true
            startedAt:
              type: string
              format: date-time
            endedAt:
              type: string
              format: date-time
              nullable: true

    CreateThreadInput:
      type: object
      required: [sessionId]
      properties:
        sessionId:
          type: string
          format: uuid
        title:
          type: string

    ThreadResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            sessionId:
              type: string
              format: uuid
            title:
              type: string
              nullable: true
            createdAt:
              type: string
              format: date-time

    ThreadSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        _count:
          type: object
          properties:
            messages:
              type: integer

    CreateMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 10000

    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        createdAt:
          type: string
          format: date-time

    PostEmotionSamplesInput:
      type: object
      required: [samples]
      properties:
        samples:
          type: array
          minItems: 1
          maxItems: 50
          items:
            type: object
            required: [source, label]
            properties:
              sessionId:
                type: string
                format: uuid
              threadId:
                type: string
                format: uuid
              timestamp:
                type: string
                format: date-time
              source:
                type: string
                enum: [self_report, webcam]
              label:
                type: string
                enum: [happy, sad, frustrated, confused, bored, neutral]
              meta:
                type: object

    EmotionSamplesResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            inserted:
              type: integer
            samples:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  label:
                    type: string
                  source:
                    type: string
                  timestamp:
                    type: string
                    format: date-time

    CreateDeckInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string

    DeckResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            title:
              type: string
            description:
              type: string
              nullable: true
            createdAt:
              type: string
              format: date-time

    DeckSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        _count:
          type: object
          properties:
            cards:
              type: integer

    CreateCardInput:
      type: object
      required: [front, back]
      properties:
        front:
          type: string
        back:
          type: string
        tags:
          type: array
          items:
            type: string

    CardResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            front:
              type: string
            back:
              type: string
            tags:
              type: array
              items:
                type: string
            createdAt:
              type: string
              format: date-time